Usernames can be picked when you join, the messages are only sent to the users in the "recepients" list, every user gets data of the others joining/leaving
For now, most of the "chosen" stuff is hardcoded, like the userName. And the users only get the messages about other users joining - but they do differentiate between joining/leaving, so it is the matter of replacing the methods called. Also probably shouldn't allow same nicknames (force difference by adding the number if it is already present?)